Hack to avoid getInitialProps being called on Safari
Safari has some issue with the back button. When a server side
navigation occurs on a web application, instead of navigating via server
side and reseting the application state when clicking on back button, it
fetchs the document via server but also keeps the application state. It
then runs the getInitialProps function which is not prepared to run on
client side.